requirements.txt
# --------- pytorch --------- #
#torch>=1.10.0
#torchvision>=0.11.0
pytorch-lightning==1.7.1
torchmetrics==0.9.3
timm
dvc
#-------- dvc-gdrive
protobuf
#----protobuf==3.20.*
#optuna
gradio
captum
shap
grad-cam
opencv-python

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
#hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
